Add 300K log lines and redater script
This script has been used in the past for workshops and demos to create 300K lines of Apache-style web logs for testing and demoing. This script will set the dates for these logs from their original time in 2014 to end at the prompted date, effectively giving a month of logs at 1K lines per day.

Both files must be placed in the root directory of a working Logstash instance. Most recently tested with 8.4.3.